Send fewer HTTP requests to control keylights
Previously, spoon.ElgatoKeys:on() was called every 1 second when
the camera was on. Now, it is only called once, when the camera
comes on.

I expect this to improve the overall system. I think I've seen
race conditions in the past, where I am trying to use Elgato's
software to change the brightness or color of the lights, but it
flickers back to an older setting, probably because my code is
constantly sending the settings back to the light based on a state
that it read.

Perhaps it would also be possible to only change the light status,
without also changing the other settings... but I haven't looked
if the API supports that.